Change ints to longs

+++ MatulaTrees/src/numbers/NumberStore.java
@@ -6,19 +6,19 @@ package numbers;
 import java.util.HashMap;
 
 public class NumberStore {
-    private static HashMap<Integer, Number> numbers = new HashMap<>();
+    private static HashMap<Long, Number> numbers = new HashMap<>();
 
-    public static void addNumber(int n) {
+    public static void addNumber(long n) {
         if (!hasNumber(n)) {
             createNumber(n);
         }
     }
 
-    private static boolean hasNumber(int n) {
+    private static boolean hasNumber(long n) {
         return numbers.containsKey(n);
     }
 
-    private static void createNumber(int n) {
+    private static void createNumber(long n) {
         Number number = new Number(n);
         numbers.put(n, number);
     }
